Update README.md changed size to limit, updated return values
I noticed that "limit" was used instead of "size" to limit the number of results returned in a query. Perhaps "size" was formerly used, or is used only in certain cases (ex. https://api.pushshift.io/reddit/search/comment/?q=trump&after=7d&aggs=created_utc&frequency=hour&size=0). But limit seems to always work, even in that case. I also observed that the default return is 100 values, and I could query and return up to 10000, I've reflected that in the table, and wherever mentioned.